refactor(*): move IngredientTypeName and IngredientWithCounter into common types
